requirements.txt
ruff==0.0.287
python-dotenv==1.0.0
currentsapi==0.0.2
boto3==1.15.8
pandas==1.3.5